\name{geesedata}
\alias{geesedata}
\docType{data}
\title{
Canada geese mark-resight data from Chapter 15
}
\description{
Spatial mark-resight dataset of 746 Canada geese (Rutledge 2012), collected in North Carolina between mid July to the end of October 2008. 
}
\usage{data(gd)}
\format{
  The format is:
List of 11
 $ y     : num [1:87, 1:21] 0 0 23 0 0 0 0 0 0 0 ...
 $ X     : num [1:87, 1:2] 3997 3988 3997 3993 3995 ...
 $ Zknown: num [1:744, 1:87, 1:21] 0 0 0 0 0 0 0 0 0 0 ...
 $ M     : num 4500
 $ EffAr : num [1:4500, 1:87, 1:21] 0 0 0 0 0 0 0 0 0 0 ...
 $ niters: num 5000
 $ xl    : num 3981
 $ xu    : num 4006
 $ yl    : num 586
 $ yu    : num 608
 $ Sex   : num [1:744] 1 2 1 1 2 2 1 1 1 1 ...
}

\source{
Rutledge, M. 2012. GET TITLE
}
\references{
Rutledge, M. 2012. GET TITLE
}

