\name{jaguarDataCh11}
\alias{jaguarDataCh11}
\docType{data}
\title{
  Jaguar data used in Chapter 9
}
\description{
  Data collected by Agustin and colleagues in Argentina.
}
\usage{data(jaguarDataCh11)}
\format{
  The format is:
List of 4
 $ jaguarCapHist   : num [1:16, 1:46, 1:96] NA NA NA NA NA NA NA NA NA NA ...
 $ jaguarStateSpace:Formal class 'RasterLayer' [package "raster"] with 16 slots
 $ jaguarPolygon   :Formal class 'SpatialPolygons' [package "sp"] with 4 slots
 $ jaguarSex       : num [1:16] 1 1 0 1 1 0 1 0 0 0 ...
}
\details{
  Some details
}
\references{
  Cite one of Agustin's papers
}
\examples{
data(jaguarDataCh11)
}
\keyword{datasets}
